Move table padding to table internal
